pyproject.toml
[project]
name = "lenslet"
version = "0.3.1"
description = "Lightweight image gallery server - no database, no files left behind"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
authors = [
    {name = "yada", email = "trojblue@gmail.com"}
]
keywords = ["image", "gallery", "server", "viewer", "photos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Viewers",
]

dependencies = [
    "fastapi>=0.135,<0.136",
    "pydantic>=2.12,<2.13",
    "uvicorn[standard]>=0.41,<0.42",
    "pillow>=12.1,<13",
    "pyarrow>=23.0,<24",
    "tqdm>=4.67,<5",
]

[project.optional-dependencies]
dev = [
    "build>=1.2",
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.6.9",
    "playwright>=1.48.0",
]
s3 = [
    "boto3>=1.34",
]
embeddings = [
    "numpy>=1.26",
]
embeddings-faiss = [
    "numpy>=1.26",
    "faiss-cpu>=1.7.4",
]

[project.scripts]
lenslet = "lenslet.cli:main"

[project.urls]
Homepage = "https://github.com/trojblue/lenslet"
Repository = "https://github.com/trojblue/lenslet"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lenslet"]
artifacts = ["src/lenslet/frontend/**"]

[tool.hatch.build.targets.sdist]
include = [
    "src/lenslet/**",
    "README.md",
    "LICENSE",
]
artifacts = ["src/lenslet/frontend/**"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests", "scripts"]
extend-exclude = [
    "src/lenslet/frontend",
]

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]
